+++ azuredevops/internal/service/core/resource_team.go
@@ -3,7 +3,6 @@ package core
 import (
 	"fmt"
 	"log"
-	"strings"
 	"time"
 
 	"github.com/ahmetb/go-linq"
@@ -339,44 +338,6 @@ func waitForTeamStateChange(d *schema.ResourceData, clients *client.AggregatedCl
 	}
 
 	return nil
-}
-
-func readTeamByName(d *schema.ResourceData, clients *client.AggregatedClient, projectID string, teamName string, top int) (*core.WebApiTeam, *schema.Set, *schema.Set, error) {
-	teamList, err := clients.CoreClient.GetTeams(clients.Ctx, core.GetTeamsArgs{
-		ProjectId:      converter.String(projectID),
-		Mine:           converter.Bool(false),
-		Top:            converter.Int(top),
-		ExpandIdentity: converter.Bool(false),
-	})
-
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	if teamList == nil || len(*teamList) <= 0 {
-		return nil, nil, nil, fmt.Errorf("Project [%s] does not contain any teams", projectID)
-	}
-
-	iTeam := linq.From(*teamList).
-		FirstWith(func(v interface{}) bool {
-			item := v.(core.WebApiTeam)
-			return strings.EqualFold(*item.Name, teamName)
-		})
-	if iTeam == nil {
-		return nil, nil, nil, fmt.Errorf("Unable to find Team with name [%s] in project with ID [%s]", teamName, projectID)
-	}
-
-	team := iTeam.(core.WebApiTeam)
-	members, err := readTeamMembers(clients, &team)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-	administrators, err := readTeamAdministrators(d, clients, &team)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	return &team, members, administrators, nil
 }
 
 func flattenTeam(d *schema.ResourceData, team *core.WebApiTeam, members *schema.Set, administrators *schema.Set) {
